Allow the config to accept custom schema and work with markdown.
Contains:
 * rework of how the plugins are build
   - diversification between HTML specific and generic plugins
   - diversification of key shortcuts between HTML specific and generic shortcuts
 * rework of how the menu is build
   - diversification of menu items between HTML specific and generic items
   - allow disabling some of the items
 * accept custom schema, if not provided then use docType specific schema
 * fix the example
