+++ react-prosemirror-config-default/src/menuItems/helpers.js
@@ -1,0 +1,41 @@
+export const markActive = type => state => {
+  const { from, $from, to, empty } = state.selection
+
+  return empty
+    ? type.isInSet(state.storedMarks || $from.marks())
+    : state.doc.rangeHasMark(from, to, type)
+}
+
+export const blockActive = (type, attrs = {}) => state => {
+  const { $from, to, node } = state.selection
+
+  if (node) {
+    return node.hasMarkup(type, attrs)
+  }
+
+  return to <= $from.end() && $from.parent.hasMarkup(type, attrs)
+}
+
+export const canInsert = type => state => {
+  const { $from } = state.selection
+
+  for (let d = $from.depth; d >= 0; d--) {
+    const index = $from.index(d)
+
+    if ($from.node(d).canReplaceWith(index, index, type)) {
+      return true
+    }
+  }
+
+  return false
+}
+
+export const promptForURL = () => {
+  let url = window && window.prompt('Enter the URL', 'https://')
+
+  if (url && !/^https?:\/\//i.test(url)) {
+    url = 'https://' + url
+  }
+
+  return url
+}
